+++ web/src/components/adminPageComponents/assistants/AdminAssistantsPersonaTable.tsx
@@ -4,7 +4,7 @@ import { Text } from "@tremor/react";
 import { Persona } from "@/interfaces/persona";
 import { useRouter } from "next/navigation";
 import { CustomCheckbox } from "@/components/CustomCheckbox";
-import { usePopup } from "@/components/adminPageComponents/connectors/Popup";
+import { usePopup } from "@/hooks/common/usePopup";
 import { useState } from "react";
 import { UniqueIdentifier } from "@dnd-kit/core";
 import { DraggableTable } from "@/components/table/DraggableTable";
