+++ web/src/components/adminPageComponents/connectors/Popup.tsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from "react";
+"use client";
 
 export interface PopupSpec {
   message: string;
@@ -14,26 +14,3 @@ export const Popup: React.FC<PopupSpec> = ({ message, type }) => (
     {message}
   </div>
 );
-
-export const usePopup = () => {
-  const [popup, setPopup] = useState<PopupSpec | null>(null);
-  // using NodeJS.Timeout because setTimeout in NodeJS returns a different type than in browsers
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
-
-  const setPopupWithExpiration = (popupSpec: PopupSpec | null) => {
-    // Clear any previous timeout
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current);
-    }
-
-    setPopup(popupSpec);
-    timeoutRef.current = setTimeout(() => {
-      setPopup(null);
-    }, 4000);
-  };
-
-  return {
-    popup: popup && <Popup {...popup} />,
-    setPopup: setPopupWithExpiration,
-  };
-};
