+++ web/src/components/adminPageComponents/documents/AdminDocumentsExplorerExplorer.tsx
@@ -7,7 +7,8 @@ import { DanswerDocument } from "@/lib/search/interfaces";
 import { buildDocumentSummaryDisplay } from "@/components/search/DocumentDisplay";
 import { CustomCheckbox } from "@/components/CustomCheckbox";
 import { updateHiddenStatus } from "@/lib/documents/helpers";
-import { PopupSpec, usePopup } from "@/components/adminPageComponents/connectors/Popup";
+import { PopupSpec } from "@/components/adminPageComponents/connectors/Popup";
+import { usePopup } from "@/hooks/common/usePopup";
 import { getErrorMsg } from "@/lib/fetchUtils";
 import { ScoreSection } from "@/components/adminPageComponents/documents/AdminDocumentsScoreEditor";
 import { useRouter } from "next/navigation";
