+++ web/src/components/adminPageComponents/models/llm/AdminModelsLLMConfiguration.tsx
@@ -7,7 +7,8 @@ import useSWR from "swr";
 import { Button, Callout, Text, Title } from "@tremor/react";
 import { ThreeDotsLoader } from "@/components/Loading";
 import { FullLLMProvider, WellKnownLLMProviderDescriptor } from "@/app/admin/models/llm/interfaces";
-import { PopupSpec, usePopup } from "@/components/adminPageComponents/connectors/Popup";
+import { PopupSpec } from "@/components/adminPageComponents/connectors/Popup";
+import { usePopup } from "@/hooks/common/usePopup";
 import { LLMProviderUpdateForm } from "./AdminModelsLLMProviderUpdateForm";
 import { LLM_PROVIDERS_ADMIN_URL } from "@/app/admin/models/llm/constants";
 import { CustomLLMProviderUpdateForm } from "./AdminModelsLLMCustomProviderUpdateForm";
