+++ web/src/components/adminPageComponents/single-connector/AdminSingleConnectorDeletionButton.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from "@tremor/react";
 import { CCPairFullInfo } from "@/app/admin/connector/[ccPairId]/types";
-import { usePopup } from "@/components/adminPageComponents/connectors/Popup";
+import { usePopup } from "@/hooks/common/usePopup";
 import { FiTrash } from "react-icons/fi";
 import { deleteCCPair } from "@/lib/documentDeletion";
 import { mutate } from "swr";
