+++ web/src/components/adminPageComponents/single-connector/AdminSingleConnectorModifyStatusButtonCluster.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from "@tremor/react";
 import { CCPairFullInfo } from "@/app/admin/connector/[ccPairId]/types";
-import { usePopup } from "@/components/adminPageComponents/connectors/Popup";
+import { usePopup } from "@/hooks/common/usePopup";
 import { disableConnector } from "@/lib/connector";
 import { mutate } from "swr";
 import { buildCCPairInfoUrl } from "@/app/admin/connector/[ccPairId]/lib";
