+++ web/src/components/chatPageComponents/folders/ChatFolderList.tsx
@@ -18,7 +18,7 @@ import {
   deleteFolder,
   updateFolderName,
 } from "./ChatFolderManagement";
-import { usePopup } from "@/components/adminPageComponents/connectors/Popup";
+import { usePopup } from "@/hooks/common/usePopup";
 import { useRouter } from "next/navigation";
 import { CHAT_SESSION_ID_KEY } from "@/lib/drag/constants";
 import Cookies from "js-cookie";
